+++ release-notes.md
@@ -3,6 +3,7 @@ RELEASE NOTES
 **Release Notes for Testy 1.8.2-SNAPSHOT**
 - add isDisplayed method in WebLocator
 - fix properties.load from inputStream correctly
+- add template for title
 
 **Release Notes for Testy 1.8.1**
 - add isDisplayed, isEnabled and submit method in WebLocatorDriverExecutor